1.3 BugFix DeathChat BugFix Groupable -> Scoreboard BugFix Historic Failure BugFix Advancement disabled BugFix Revive Typo Added: Random, and no team in TeamChooseGUI Rework: GUI

+++ src/fr/supercomete/head/Listeners/OnExpGainListener.java
@@ -4,12 +4,11 @@ import fr.supercomete.head.GameUtils.GameConfigurable.Configurable;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerExpChangeEvent;
-
 import fr.supercomete.head.core.Main;
 
 final class OnExpGainListener implements Listener {
 	@EventHandler
 	public void onExpGain(PlayerExpChangeEvent e) {
-		e.setAmount(e.getAmount()*(Main.currentGame.getDataFrom(Configurable.LIST.ExpBoost)/100));
+		e.setAmount((int)((double)e.getAmount()*((double)Main.currentGame.getDataFrom(Configurable.LIST.ExpBoost)/100.0)));
 	}
 }
